Connect 4 Game Completion
The game is complete, it features a turn based system where you select which column you want to drop your coin in. The column will also indicate whos turn it is and show where it will land when hovering over the column. Audio implemented and auto detection of a winner horizontally, diagonally, and vertically.